refactor: move createSessionsFromPrekeys to sessionhandler
